Update emojify.js
Removed "navigator.clipboard.writeText(final);" because it would only work intermittently. Perhaps more importantly, when it did work, it would ask the user to give copy-to-clipboard permissions to whatever website the browser was on when the bookmark script was run, and I'm not certain I want to train users to get accustomed to giving all the websites copy-to-clipboard permissions.